:sparkles: (optionsScreen) show options in a dialog instead of a screen
