+++ lib/main.dart
@@ -5,7 +5,6 @@ import 'package:snake_game/screens/game.dart';
 import 'package:snake_game/screens/landing.dart';
 import 'package:snake_game/screens/levels.dart';
 import 'package:snake_game/screens/login.dart';
-import 'package:snake_game/screens/options.dart';
 
 void main() => runApp(MyApp());
 
@@ -23,7 +22,6 @@ class MyApp extends StatelessWidget {
         routes: {
           "/": (context) => Landing(),
           "/game": (context) => Game(),
-          "/options": (context) => Options(),
           "/login": (context) => Login(),
           "/levels": (context) => Levels(),
         },
